moved everything LNbits related to own file, cleanup.
